Add repo type and repo_gpgcheck settings to repo files
Zypper requires the repository type to be declared in repo files.
According to zypper(8), standard RPM MetaData repositories are declared
as "type=rpm-md".

Zypper also defaults to assuming that repodata is signed when "gpgcheck=1"
is set in the repo file. Since RPM Fusion does not currently sign the repodata,
this needs to be declared with "repo_gpgcheck=0".